pyproject.toml
[tool.poetry]
name = "hw1-ml-fastapi"
version = "0.1.0"
description = ""
authors = ["DK <d.kachkin@tnxholding.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
orjson = "^3.9.10"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
ecs-logging = "^2.1.0"
celery = "^5.3.5"
uvicorn = "^0.24.0.post1"
redis = "^5.0.1"
scikit-learn = "^1.3.2"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"


[tool.poetry.dev-dependencies]
mypy = "^1.7.0"
flake8 = "^6.1.0"
black = { extras = ["jupyter"], version = "^23.11.0" }
pandas = "^2.1.3"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
jupyter = "^1.0.0"
mlflow = "^2.8.1"
psycopg2-binary = "^2.9"
boto3 = "^1.29.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | venv
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
multi_line_output = 3
